Added Password Authentication and Updated Pages
Implemented password authentication functionality and separated the to-do list and timer to separate pages. Created new file for the main authenticated page.

This way, users will be prompted to enter the password on the homepage, and after successful authentication, they will be redirected to the authenticated.html page where they can access the utilities and games from the navbar.